backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.29.0
motor==3.6.0         # async MongoDB driver
python-jose[cryptography]==3.3.0   # JWT
bcrypt==4.1.3
python-dotenv==1.0.1
pydantic[email]==2.10.6
pymongo==4.9.2